+++ sources/System/Config.cpp
@@ -46,7 +46,7 @@ void		Game::Config::initialize(int argc, char** argv)
 #endif
 
   // Initialize random
-  std::srand(std::time(nullptr));
+  std::srand((unsigned int)std::time(nullptr));
 
   // Detect maximum of thread concurrency.
   Game::Config::ThreadNumber = std::thread::hardware_concurrency();
